requirements.txt
langchain 
faiss-cpu
langchain-community
langchain-google-genai
ipykernel
Flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.2.0